Somativa/Pecas/requirements.txt
flask==2.1.3
requests==2.31.0
Werkzeug==2.0.3
mysql-connector-python